+++ src/main/java/bteam/example/ecoolshop/controller/GoodsController.java
@@ -4,8 +4,8 @@ import bteam.example.ecoolshop.dto.GoodsDto;
 import bteam.example.ecoolshop.entity.Goods;
 import bteam.example.ecoolshop.service.GoodsService;
 import bteam.example.ecoolshop.util.ConversationUtil;
-import bteam.example.ecoolshop.util.Order;
-import bteam.example.ecoolshop.util.ResponseMap;
+import bteam.example.ecoolshop.util.enums.Order;
+import bteam.example.ecoolshop.util.maps.ResponseMap;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
