+++ src/main/java/bteam/example/ecoolshop/util/enums/SecurityRoles.java
@@ -1,4 +1,4 @@
-package bteam.example.ecoolshop.util;
+package bteam.example.ecoolshop.util.enums;
 
 public interface SecurityRoles {
     String REGULAR = "REGULAR_USER";
